Update of generated kernels

+++ XpulpNN-mixed/32bit/src/Convolution/xpulp_nn_mix_conv_u2_u2_i2.c
@@ -240,7 +240,7 @@ void __attribute__((noinline)) xpulp_nn_mix_conv_u2_u2_i2(
 
         for(int j=0; j < (num_col_im2col >> 4); j++)
         {
-          sum = SumDotp16(*(uint8_t *)ptrB, *(int32_t *)ptrA, sum);
+          sum = SumDotp16(*(uint32_t *)ptrB, *(int32_t *)ptrA, sum);
           ptrA++;
           ptrB++;
         }
