+++ XpulpNN-mixed/32bit/src/Convolution/xpulp_nn_mix_conv_u2_u4_i4.c
@@ -234,7 +234,7 @@ void __attribute__((noinline)) xpulp_nn_mix_conv_u2_u4_i4(
 
         for(int j=0; j < (num_col_im2col >> 3); j++)
         {
-          sum = SumDotp8(*(uint8_t *)ptrB, *(int32_t *)ptrA, sum);
+          sum = SumDotp8(*(uint32_t *)ptrB, *(int32_t *)ptrA, sum);
           ptrA++;
           ptrB++;
         }
