+++ XpulpNN-mixed/32bit/src/Convolution/xpulp_nn_mix_conv_u4_u8_i2_4x4.c
@@ -235,11 +235,11 @@ void __attribute__((noinline)) xpulp_nn_mix_conv_u4_u8_i2_4x4(
 
         for(int j=0; j < (num_col_im2col >> 4); j++)
         {
-          sum = SumDotp8(*(uint8_t *)ptrB, *(int32_t *)ptrA, sum);
+          sum = SumDotp8(*(uint32_t *)ptrB, *(int32_t *)ptrA, sum);
 
           ptrB++;
 
-          sum = SumDotp8(*(uint8_t *)ptrB, *(int32_t *)ptrA, sum);
+          sum = SumDotp8(*(uint32_t *)ptrB, *(int32_t *)ptrA, sum);
 
           ptrA++;
           ptrB++;
